+++ Retia.Gui/Models/TrainOptionsModel.cs
@@ -10,7 +10,7 @@ namespace Retia.Gui.Models
         public int ErrorFilterSize { get; set; } = 100;
         public float LearningRate { get; set; } = 0.0001f;
         public int LearningRateScalePeriod { get; set; } = 50;
-        public double LearningRateScaleFactor { get; set; } = 0.01f;
+        public float LearningRateScaleFactor { get; set; } = 0.01f;
         public int MaxEpoch { get; set; } = 100000;
 
         public RelayCommand StartResumeCommand { get; set; }
