+++ Retia/Training/Trainers/TrainReportEventArgsBase.cs
@@ -4,13 +4,13 @@ namespace Retia.Training.Trainers
 {
     public abstract class TrainReportEventArgsBase : EventArgs
     {
-        protected TrainReportEventArgsBase(double epoch, long iteration)
+        protected TrainReportEventArgsBase(long epoch, long iteration)
         {
             Epoch = epoch;
             Iteration = iteration;
         }
 
-        public double Epoch { get; set; }
+        public long Epoch { get; set; }
         public long Iteration { get; set; }
     }
 }
